Move all other endpoints as well
